Create Week #4 Activity Project
I ultimately ended up predicting the classe in the test set using Random Forest after testing that method as well as rpart and SVM.  Cross Validation was performed by taking the Training set (dat_train) and partitioning it into a training (training) and a test (testing) set (75%/25%).  The models were fitted on the training dataset, and tested on the testing dataset. Once it was established that Random Forest was the most accurate, that method was tested on the original Testing (dat_test) data set.  After I partitioned the training set, I then weeded out the NZV variables to condense my columns and ensure I was using meaningful predictors.  This took my variables down from well over 160 to 58 (including the Classe variable which I removed since that was what I was predicting).  
The out-of-sample error will correspond to the expected number of mis-classified observations/total observations in the Test data set.  The expected out-of-sample error will correspond to the quantity: 1-accuracy in the cross-validation data. Accuracy is the % of correct classified observations over the total sample in the partitioned test data set, which provided the expected accuracy.  Based on the Random Forest showing 99.98% accuracy, I would expect to be wrong 0.02 times in the 20 predictions in the test set.  As seen from the above, Random Forest was the most logical choice based on it's accuracy %.